requirements.txt
google-genai
langsmith
python-dotenv
numpy
scipy
matplotlib